\name{Sarsa-package}
\alias{Sarsa-package}
\alias{Sarsa}
\docType{package}
\title{
\packageTitle{Sarsa}
}
\description{
\packageDescription{Sarsa}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Sarsa}
\packageIndices{Sarsa}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Sarsa}

Maintainer: \packageMaintainer{Sarsa}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
